Refactor package structure for note activities
