+++ app/src/main/AndroidManifest.xml
@@ -56,12 +56,12 @@
             android:screenOrientation="locked"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".activity.note.NotesListActivity"
+            android:name=".activity.note.list.NotesListActivity"
             android:exported="false"
             android:screenOrientation="locked"
             android:theme="@style/AppTheme" />
         <activity
-            android:name=".activity.note.OpenNoteActivity"
+            android:name=".activity.note.editor.OpenNoteActivity"
             android:exported="false"
             android:screenOrientation="locked"
             android:theme="@style/AppTheme" />
