+++ app/src/main/java/app/notesr/activity/importer/ImportActivity.java
@@ -28,7 +28,7 @@ import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import app.notesr.R;
 import app.notesr.activity.ActivityBase;
-import app.notesr.activity.note.NotesListActivity;
+import app.notesr.activity.note.list.NotesListActivity;
 import app.notesr.service.AndroidServiceRegistry;
 import app.notesr.service.importer.ImportAndroidService;
 import app.notesr.service.importer.ImportAndroidServiceStarter;
