+++ app/src/main/java/app/notesr/activity/note/editor/OpenFilesListAction.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-package app.notesr.activity.note;
+package app.notesr.activity.note.editor;
 
 import android.content.Intent;
