+++ app/src/main/java/app/notesr/activity/note/list/OpenNoteOnClick.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-package app.notesr.activity.note;
+package app.notesr.activity.note.list;
 
 import android.content.Context;
 import android.content.Intent;
@@ -12,6 +12,7 @@ import android.widget.AdapterView;
 
 import java.util.Map;
 
+import app.notesr.activity.note.editor.OpenNoteActivity;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
