+++ app/src/main/java/app/notesr/activity/security/ReEncryptionActivity.java
@@ -18,7 +18,7 @@ import java.nio.charset.CharacterCodingException;
 import app.notesr.R;
 import app.notesr.activity.ActivityBase;
 import app.notesr.activity.DialogFactory;
-import app.notesr.activity.note.NotesListActivity;
+import app.notesr.activity.note.list.NotesListActivity;
 import app.notesr.service.AndroidServiceRegistry;
 import app.notesr.service.security.crypto.update.SecretsUpdateAndroidService;
 import app.notesr.service.security.crypto.update.SecretsUpdateAndroidServiceStarter;
