+++ app/src/main/res/layout/activity_note_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/activity_background"
-    tools:context=".activity.note.NotesListActivity">
+    tools:context=".activity.note.list.NotesListActivity">
 
     <TextView
         android:id="@+id/missingNotesLabel"
