+++ app/src/main/res/layout/activity_open_note.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/activity_background"
-    tools:context=".activity.note.OpenNoteActivity">
+    tools:context=".activity.note.editor.OpenNoteActivity">
 
     <EditText
         android:id="@+id/noteNameField"
